Backtest build block app for synthetic blocks. (#344)
## 📝 Summary

This PR makes backtest-build-block more versatile so the chain used and
orders come from an ```OrdersSource``` instead of the hardcoded mainnet
we had.
The original backtest_build_block behavior is now achieved using the
```OrdersSource``` ```LandedBlockFromDBOrdersSource```.
A new ```OrdersSource``` ```SyntheticOrdersSource``` is provided as an
example of how to run backtest-build-block with a syntetic chain and
synthetic orders.


## 💡 Motivation and Context

<!--- (Optional) Why is this change required? What problem does it
solve? Remove this section if not applicable. -->

---

## ✅ I have completed the following steps:

* [X] Run `make lint`
* [X] Run `make test`
* [ ] Added tests (if applicable)

---------

Co-authored-by: sukoneck <19413126+sukoneck@users.noreply.github.com>